tests/test_cli.py
from click.testing import CliRunner
from app.cli import import_data, export_data

def test_data_import_export():
    runner = CliRunner()

    # Provide a test CSV file path for import
    import_file_path = 'test_import.csv'
    
    # Simulate the import command-line invocation
    import_result = runner.invoke(import_data, [import_file_path])
    
    # Assert the expected output or behavior for import
    assert import_result.exit_code == 0
    assert import_result.output.strip() == "Data import successful!"
    
    # Provide a test CSV file path for export
    export_file_path = 'test_export.csv'
    
    # Simulate the export command-line invocation
    export_result = runner.invoke(export_data, [export_file_path])
    
    # Assert the expected output or behavior for export
    assert export_result.exit_code == 0
    assert export_result.output.strip() == "Data export successful!"

    # Add additional assertions for the imported and exported data or files, if needed

# Add more test cases as needed

if __name__ == '__main__':
    test_data_import_export()
